+++ contracts/misc/BTCZapBsc.sol
@@ -40,17 +40,17 @@ contract BTCZapBsc {
 
     constructor() public {
         governance = msg.sender;
-        IERC20Upgradeable(BTCB).safeApprove(VENUS_BTC, uint256(-1));
-        IERC20Upgradeable(VENUS_BTC).safeApprove(VENUS_BTC_PLUS, uint256(-1));
+        IERC20Upgradeable(BTCB).safeApprove(VENUS_BTC, uint256(int256(-1)));
+        IERC20Upgradeable(VENUS_BTC).safeApprove(VENUS_BTC_PLUS, uint256(int256(-1)));
 
-        IERC20Upgradeable(BTCB).safeApprove(FORTUBE_CONTROLLER, uint256(-1));
-        IERC20Upgradeable(FORTUBE_BTCB).safeApprove(FORTUBE_BTCB_PLUS, uint256(-1));
+        IERC20Upgradeable(BTCB).safeApprove(FORTUBE_CONTROLLER, uint256(int256(-1)));
+        IERC20Upgradeable(FORTUBE_BTCB).safeApprove(FORTUBE_BTCB_PLUS, uint256(int256(-1)));
 
-        IERC20Upgradeable(BTCB).safeApprove(ACS_BTCB, uint256(-1));
-        IERC20Upgradeable(ACS_BTCB).safeApprove(ACS_BTCB_PLUS, uint256(-1));
+        IERC20Upgradeable(BTCB).safeApprove(ACS_BTCB, uint256(int256(-1)));
+        IERC20Upgradeable(ACS_BTCB).safeApprove(ACS_BTCB_PLUS, uint256(int256(-1)));
 
-        IERC20Upgradeable(BTCB).safeApprove(AUTO_BTC, uint256(-1));
-        IERC20Upgradeable(AUTO_BTC).safeApprove(AUTO_BTC_PLUS, uint256(-1));
+        IERC20Upgradeable(BTCB).safeApprove(AUTO_BTC, uint256(int256(-1)));
+        IERC20Upgradeable(AUTO_BTC).safeApprove(AUTO_BTC_PLUS, uint256(int256(-1)));
     }
 
     /**
